Rename: 커뮤니티 내부 컴포넌트별 폴더 삭제
