+++ src/components/community/CommunityItem.tsx
